Create surveyCreatorConfig file
